Updating ArchUnit classes structure
